HomeQuizServer/requirements.txt
flask>=2.0.0
flask-socketio>=5.1.0
flask-cors>=3.0.0
pymongo>=4.0.0
tinydb>=4.7.0
python-dotenv>=0.19.0
cloudinary>=1.28.0
requests>=2.26.0
pyinstaller>=4.5.0
